try and cleanup percell timing (#248)

+++ Core/Inc/can_messages.h
@@ -110,7 +110,7 @@ void send_segment_temp_message(acc_data_t *bmsdata);
 void send_fault_message(uint8_t status, int16_t curr, int16_t in_dcl);
 
 void send_fault_timer_message(uint8_t start_stop, uint32_t fault_code,
-			      uint16_t data_1);
+			      float data_1);
 
 /**
  * @brief Send CAN message for debugging the car on the fly.
